requirements.txt
flask
flask-login
flask-sqlalchemy
werkzeug
pillow